IOT-HOME AUTOMATION
Home Automation (IoT)- Winner
Feb 2020-A contest by Cutting Edge visinories,SVNIT

Project descriptionInternet is making our life easy in a lot many ways. One of them was on which we were working on. It is a very basic model of Home Automation with the help of Internet. Basic components used for this purpose were:
1.Relay Board
2.Node MCU
3.Android Application using firebase